refactor and add doc to tests
